Use ``InteractiveReporter`` in Presentation Compiler tests
By using the ``InteractiveReporter``, all compiler's errors occurring in a
compilation unit are now collected in the compilation unit itself. This was
necessary for testing SI-5975.

You may wonder why compilation's errors are not directly reported in the test's
output.  The reason is that often you want to make sure that semantic actions
(such as code completion, hyperlinking, and similar) are available despite the
unit may contain some compilation errors. Indeed, many of the existing
Presentation Compiler tests have this property.