fix SI-5975
what happened:
1. first compiler run creates a module symbol of a case-companion with a lazy type. the case class has a default argument
2. nobody looks at it, so the type doesn't get completed
3. in a second compiler run, the type completion is triggered

the problem was that `resetNamer` would clear the `classAndNamerOfModule` map, which is too eager: the type completer still needs entries from there. fixed by turning it into a weak map.

No longer storing the namer of the companion object in the HashMap, but as a tree attachment instead. Otherwise, GC will delete the namer from the weak map. As an attachment, it gets collected together with the ClassDef tree node.